Move FEN parser into separate module
